Second thoughs on AO visimap bitmap format.
AO visimap is kept in a Bitmapset, and materialized to disk as a bitmap
consisting of 32-bit words. But Bitmapset uses 64-bit words in upstream
now. Need to convert. Or change the bitmap compression routines to deal
with 64-bit words. I'm not sure which is better.

+++ src/backend/utils/misc/bitmap_compression.c
@@ -86,7 +86,7 @@ BitmapDecompress_HasError(
  */
 void
 BitmapDecompress_Decompress(BitmapDecompressState *state,
-							uint64 *bitmap,
+							uint32 *bitmap,
 							int bitmapDataSize)
 {
 	uint32 lastBlockData, flag, rleRepeatCount;
